requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alpaca-trade-api==1.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.9.3
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backtrader==1.9.76.123
black==22.1.0
bleach==4.1.0
blinker==1.4
blis==0.7.5
Bottleneck==1.3.2
brbot==0.1.0
brotlipy==0.7.0
CacheControl==0.12.10
cachetools==4.2.4
cachy==0.3.0
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
cleo==0.8.1
click==8.0.4
click-completion==0.5.2
click-default-group==1.2.2
clikit==0.6.2
colorama==0.4.4
commonmark==0.9.1
crashtest==0.3.1
cryptography==36.0.0
cucumber-tag-expressions==4.1.0
cycler==0.11.0
cyclonedx-python-lib==1.3.0
cymem==2.0.6
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.4
dodgy==0.2.1
entrypoints==0.3
filelock==3.6.0
flake8==4.0.1
flake8-polyfill==1.0.2
Flask==2.0.3
fonttools==4.25.0
frozenlist==1.3.0
ghp-import==2.0.2
gitdb==4.0.9
GitPython==3.1.27
google-api-core==1.31.5
google-auth==1.33.0
google-auth-oauthlib==0.5.0
google-cloud-bigquery==1.10.0
google-cloud-core==0.29.1
google-crc32c==1.1.2
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
graphviz==0.19.1
greenlet==1.1.2
grpcio==1.43.0
html5lib==1.1
htmlmin==0.1.12
idna==3.3
ImageHash==4.2.1
importlib-metadata==4.11.2
ipykernel==6.4.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.0.1
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==23.5.0
kiwisolver==1.3.1
langcodes==3.3.0
lazy-object-proxy==1.7.1
llvmlite==0.38.0
lockfile==0.12.2
lxml==4.8.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mergedeep==1.3.4
MetaTrader5==5.0.36
missingno==0.5.0
mistune==0.8.4
mkdocs==1.2.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
msgpack==1.0.3
multidict==6.0.2
multimethod==1.7
multitasking==0.0.10
munkres==1.1.4
murmurhash==1.0.6
mypy-extensions==0.4.3
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
nltk==3.7
notebook==6.4.8
numba==0.55.0
numexpr==2.8.1
numpy==1.22.2
oauthlib==3.2.0
olefile==0.46
packageurl-python==0.9.9
packaging==20.9
pandas==1.4.1
pandas-datareader==0.10.0
pandas-gbq==0.10.0
pandas-ta==0.3.14b0
pandocfilters==1.5.0
parso==0.8.3
pastel==0.2.1
pathspec==0.9.0
pathy==0.6.1
patsy==0.5.2
pep8-naming==0.10.0
pexpect==4.8.0
phik==0.11.2
pickleshare==0.7.5
Pillow==8.4.0
pip==21.2.4
pip-api==0.0.28
pip-audit==2.0.0
pkginfo==1.8.2
platformdirs==2.5.1
plotly==5.6.0
pluggy==1.0.0
poetry-core==1.0.8
pprintpp==0.4.0
preshed==3.0.6
progress==1.6
prometheus-client==0.13.1
prometheus-flask-exporter==0.18.7
prompt-toolkit==3.0.20
prospector==1.7.5
protobuf==3.19.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pydata-google-auth==1.3.0
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==2.3.0
pylev==1.4.0
pylint==2.12.2
pylint-celery==0.3
pylint-django==2.5.2
pylint-flask==0.6
pylint-plugin-utils==0.7
pymarketstore==0.20
pynndescent==0.5.6
pyOpenSSL==22.0.0
pyparsing==3.0.7
pypiwin32==223
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pyu2f==0.1.5
PyWavelets==1.2.0
pywin32==302
pywin32-ctypes==0.2.0
pywinpty==2.0.2
PyYAML==6.0
pyyaml_env_tag==0.1
pyzmq==22.3.0
querystring-parser==1.2.4
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requirements-detector==0.7
resolvelib==0.8.1
rich==10.16.2
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
setoptconf-tmp==0.3.1
setuptools==58.0.4
shellingham==1.4.0
sip==4.19.13
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
snowballstemmer==2.2.0
spacy-legacy==3.0.9
spacy-loggers==1.0.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
srsly==2.4.2
statsmodels==0.13.2
tabulate==0.8.9
tangled-up-in-unicode==0.1.0
tenacity==8.0.1
terminado==0.13.1
testpath==0.6.0
thinc==8.0.13
threadpoolctl==2.2.0
toml==0.10.2
tomli==1.2.3
tomlkit==0.10.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typer==0.4.0
types-setuptools==57.4.9
types-toml==0.10.4
typing_extensions==4.1.1
urllib3==1.26.8
virtualenv==20.13.2
visions==0.7.4
waitress==2.0.0
ward==0.65.0b0
wasabi==0.9.0
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
websockets==9.1
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.5.2
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.13.3
yarl==1.7.2
yfinance==0.1.70
zipp==3.7.0
